docs(changeset): Minor improvements
- Type definitions moved to separate file
- `scrollContainerRef` ref handle now receives ref, not DOM node
