+++ .changeset/clean-ghosts-cry.md
@@ -1,0 +1,8 @@
+---
+'@wldyslw/react-bottom-sheet': patch
+---
+
+Minor improvements
+
+-   Type definitions moved to separate file
+-   `scrollContainerRef` ref handle now receives ref, not DOM node
