+++ packages/react-bottom-sheet/src/BottomSheet/index.ts
@@ -1,0 +1,4 @@
+import BottomSheet from './BottomSheet';
+
+export type { BottomSheetProps, BottomSheetRef } from './BottomSheet.types';
+export default BottomSheet;
