--- packages/react-bottom-sheet/src/BottomSheet/main.ts
@@ -1,4 +1,0 @@
-import BottomSheet from './BottomSheet';
-
-export type { BottomSheetProps, BottomSheetRef } from './BottomSheet';
-export default BottomSheet;
